refactor LeaseAssignmentDecider back to interface
